+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Telegram Bot API for PHP Change Log
 
-## 0.7.1 under development
+## 0.7.1 February 12, 2025
 
 - New #142: Add `CurlTransport`.
 - New #144: Add `chatId` parameter to `SendGift` method.